Created welcome component and added styles for drop down
