+++ src/Locations.css
@@ -1,11 +1,44 @@
-body {
- background: url(./images/splash-background.jpeg) no-repeat center center fixed;
- -webkit-background-size: cover;
- -moz-background-size: cover;
- -o-background-size: cover;
- background-size: cover;
-} 
-
-.splash-page {
- 
+* {
+  box-sizing: border-box;
 }
+
+.location-container {
+  width: 500px;
+}
+
+
+h2 {
+  padding: 20px 0;
+  margin: 0;
+  text-align: center;
+  background: rgba(255, 255, 255, 0.75);
+  border-radius: 5px 5px 0 0;
+  border-bottom: .5px solid gray;
+}
+
+h2:hover {
+  background: rgba(250, 209, 125, 0.75);
+}
+
+ul {
+  list-style-type: none;
+  position: absolute;
+  height: 35%;
+  overflow: auto;
+  padding: 0;
+  margin: 0;
+  border-radius: 0 0 5px 5px;
+}
+
+li {
+  font-size: 24px;
+  padding: 10px;
+  width: 500px;
+  background: rgba(255, 255, 255, 0.75);
+  border-bottom: 1px solid gray;
+}
+
+li:hover {
+  background: rgba(250, 209, 125, 0.75);
+}
+
