+++ src/Welcome.css
@@ -1,0 +1,11 @@
+body {
+ background: url(./images/splash-background.jpeg) no-repeat center center fixed;
+ -webkit-background-size: cover;
+ -moz-background-size: cover;
+ -o-background-size: cover;
+ background-size: cover;
+ display: flex;
+ justify-content: center;
+ align-items: center;
+ height: 100vh;
+}
